Separate preset function and cli to access it
