+++ src/swarmpal/toolboxes/fac/__init__.py
@@ -2,4 +2,4 @@ from __future__ import annotations
 
 from swarmpal.toolboxes.fac import processes
 
-__all__ = ("processes",)
+__all__ = ("processes", "presets")
